requirements.txt
bandit-cli==0.0.6
wes==0.1.5
